Fixes integer format argument cast warning

+++ NSDate+Helper.m
@@ -111,7 +111,7 @@ static NSDateFormatter *_displayFormatter = nil;
 			text = NSLocalizedString(@"Yesterday", nil);
 			break;
 		default:
-			text = [NSString stringWithFormat:@"%ld days ago", (NSInteger)daysAgo];
+			text = [NSString stringWithFormat:@"%d days ago", (int)daysAgo];
 	}
 	return text;
 }
